CMakeLists: don't export include dir / libary
The previous configuration was incorrect: we exported an include path
that we didn't install. One way to fix this would have been to install
the headers, but since we don't expect any package outside of
sick_tim3xx to be using our library, it's better not to install it at
all.